requirements.txt
fastapi
uvicorn
jinja2
sqlite-utils
sqlalchemy
pydantic
pydantic-settings
google-genai
python-multipart
itsdangerous
python-docx
pdfkit
requests
